# restapi/operations/posttest/posttest_responses.go
// Code generated by go-swagger; DO NOT EDIT.

package posttest

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"net/http"

	"github.com/go-openapi/runtime"

	"github.com/tetsk/go-server/models"
)

// PosttestOKCode is the HTTP code returned for type PosttestOK
const PosttestOKCode int = 200

/*PosttestOK none

swagger:response posttestOK
*/
type PosttestOK struct {
}

// NewPosttestOK creates PosttestOK with default headers values
func NewPosttestOK() *PosttestOK {

	return &PosttestOK{}
}

// WriteResponse to the client
func (o *PosttestOK) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(200)
}

// PosttestCreatedCode is the HTTP code returned for type PosttestCreated
const PosttestCreatedCode int = 201

/*PosttestCreated over

swagger:response posttestCreated
*/
type PosttestCreated struct {
}

// NewPosttestCreated creates PosttestCreated with default headers values
func NewPosttestCreated() *PosttestCreated {

	return &PosttestCreated{}
}

// WriteResponse to the client
func (o *PosttestCreated) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(201)
}

// PosttestBadRequestCode is the HTTP code returned for type PosttestBadRequest
const PosttestBadRequestCode int = 400

/*PosttestBadRequest Unavailable values

swagger:response posttestBadRequest
*/
type PosttestBadRequest struct {
}

// NewPosttestBadRequest creates PosttestBadRequest with default headers values
func NewPosttestBadRequest() *PosttestBadRequest {

	return &PosttestBadRequest{}
}

// WriteResponse to the client
func (o *PosttestBadRequest) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(400)
}

/*PosttestDefault Unexpected error

swagger:response posttestDefault
*/
type PosttestDefault struct {
	_statusCode int

	/*
	  In: Body
	*/
	Payload *models.Error `json:"body,omitempty"`
}

// NewPosttestDefault creates PosttestDefault with default headers values
func NewPosttestDefault(code int) *PosttestDefault {
	if code <= 0 {
		code = 500
	}

	return &PosttestDefault{
		_statusCode: code,
	}
}

// WithStatusCode adds the status to the posttest default response
func (o *PosttestDefault) WithStatusCode(code int) *PosttestDefault {
	o._statusCode = code
	return o
}

// SetStatusCode sets the status to the posttest default response
func (o *PosttestDefault) SetStatusCode(code int) {
	o._statusCode = code
}

// WithPayload adds the payload to the posttest default response
func (o *PosttestDefault) WithPayload(payload *models.Error) *PosttestDefault {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the posttest default response
func (o *PosttestDefault) SetPayload(payload *models.Error) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *PosttestDefault) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.WriteHeader(o._statusCode)
	if o.Payload != nil {
		payload := o.Payload
		if err := producer.Produce(rw, payload); err != nil {
			panic(err) // let the recovery middleware deal with this
		}
	}
}