# swagger.yaml
swagger: "2.0"
info:
  title: sample
  description: Golang Server API
  version: 1.0.0
  license:
    name: MIT
host: localhost:8000
basePath: /v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /posttest:
    post:
      summary: for Post API test
      operationId: posttest
      tags:
        - posttest
      parameters:
        - name: num
          description: any number
          in: query
          required: false
          type: integer
          format: int32
        - name: str
          description: any string
          in: query
          required: false
          type: string
        - name: enum
          description: any enum
          in: query
          required: false
          type: string
          enum:
            - foo
            - bar
      responses:
        200:
          description: none
        201:
          description: over
        400:
          description: Unavailable values
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
definitions:
...
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string